[#1274] clean up Mtx data source
